requirements.txt
torch==2.5.0
torchvision==0.20.0
diffusers==0.33.0
transformers==4.49.0
Pillow==11.0.0
numpy==1.26.4
opencv-python==4.11.0.86
accelerate==0.30.1
gradio==5.23.3
protobuf==5.28.3
sentencepiece==0.2.0
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
aiohttp==3.9.5